Deno/Rust: Fibonacci recursivo

+++ src/deno_ffi/algorithm.rs
@@ -9,7 +9,7 @@ pub extern "C" fn sub(a: isize, b: isize) -> isize {
 }
 
 #[no_mangle]
-pub extern "C" fn fib(n: isize) -> isize {
+pub extern "C" fn fib(n: u64) -> u64 {
     match n {
         0 | 1 => n,
         _ => fib(n - 1) + fib(n - 2),
